Refactor user to Customer
